refactored parsing event objects
